Minimally usable (#3)
* main initial implementation

* add long decode test

* index and view

* read from file and write to file + encode decode test

* update makefile and workflow

* update windows makefile

* enable optimization for release

* release workflow

* bumpconfig and python code quality

* setup.cfg

* rename fastq to example.fastq and include in release

* update usage strings

+++ src/reads.cpp
@@ -44,7 +44,7 @@ size_t Reads::read_n_bases() {
     throw std::runtime_error("Malformed n_bases");
   }
   for (uint8_t i = 0; i < this->BASES_BYTES; i++) {
-    n_bases += n_bases_buffer[i]
+    n_bases += ((size_t)n_bases_buffer[i])
                << (Sequence::TYPE_SIZE * (this->BASES_BYTES - i - 1));
   }
   delete[] n_bases_buffer;
